requirements.txt
opencv-python
mediapipe==0.10.9
protobuf==3.20.3
scikit-learn
pandas
numpy<2.0
pillow
tqdm